+++ .github/workflows/deploy-dev.yml
@@ -41,7 +41,6 @@ jobs:
       - name: Publish KV to Development
         run: |
           npm run kv-causes:development
-          npm run kv-conversion-rate:development
           npm run kv-summary:development
           npm run kv-trends-previous:development
         working-directory: ./workers/tiltify-cache
