+++ .github/workflows/deploy-production.yml
@@ -42,7 +42,6 @@ jobs:
       - name: Publish KV to Production
         run: |
           npm run kv-causes:production
-          npm run kv-conversion-rate:production
           npm run kv-summary:production
           npm run kv-trends-previous:production
         working-directory: ./workers/tiltify-cache
