+++ website/home.html
@@ -6,6 +6,10 @@
         
         <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
         <meta name="viewport" content="initial-scale=1">
+
+        <link rel="icon" type="image/x-icon" href="/assets/favicon.ico">
+
+        <title>Jingle Jam Tracker</title>
     </head>
     <body>
         <div id='myContainer'></div>
