--- workers/tiltify-cache/kv/conversion-rate.txt
@@ -1,1 +1,0 @@
-1.245
