+++ workers/tiltify-cache/package.json
@@ -9,15 +9,12 @@
       "deploy:production": "wrangler deploy",
       "deploy:development": "wrangler deploy -e development",
       "kv-causes:production": "wrangler kv:key put causes --path=kv/causes.json --namespace-id=9d285e05684c4887acb7731de7a41a4e",
-      "kv-conversion-rate:production": "wrangler kv:key put conversion-rate --path=kv/conversion-rate.txt --namespace-id=9d285e05684c4887acb7731de7a41a4e",
       "kv-summary:production": "wrangler kv:key put summary --path=kv/summary.json --namespace-id=9d285e05684c4887acb7731de7a41a4e",
       "kv-trends-previous:production": "wrangler kv:key put trends-previous --path=kv/trends-previous.json --namespace-id=9d285e05684c4887acb7731de7a41a4e",
       "kv-causes:development": "wrangler kv:key put causes --path=kv/causes.json --namespace-id=c9634f49deeb4618b3c3f1980f777f7a",
-      "kv-conversion-rate:development": "wrangler kv:key put conversion-rate --path=kv/conversion-rate.txt --namespace-id=c9634f49deeb4618b3c3f1980f777f7a",
       "kv-summary:development": "wrangler kv:key put summary --path=kv/summary.json --namespace-id=c9634f49deeb4618b3c3f1980f777f7a",
       "kv-trends-previous:development": "wrangler kv:key put trends-previous --path=kv/trends-previous.json --namespace-id=c9634f49deeb4618b3c3f1980f777f7a",
       "kv-causes:local": "wrangler kv:key put causes --path=kv/causes.json --namespace-id=6a08d74af525475baefa19a93f0ba136 --preview --local",
-      "kv-conversion-rate:local": "wrangler kv:key put conversion-rate --path=kv/conversion-rate.txt --namespace-id=6a08d74af525475baefa19a93f0ba136 --preview --local",
       "kv-summary:local": "wrangler kv:key put summary --path=kv/summary.json --namespace-id=6a08d74af525475baefa19a93f0ba136 --preview --local",
       "kv-trends-previous:local": "wrangler kv:key put trends-previous --path=kv/trends-previous.json --namespace-id=6a08d74af525475baefa19a93f0ba136 --preview --local"
     },
