+++ workers/tiltify-cache/wrangler.toml
@@ -24,6 +24,7 @@ YEAR=2023
 COLLECTIONS_AVAILABLE=100000
 DOLLAR_OFFSET=3371741.16
 DONATION_DIFFERENCE=0
+CONVERSION_RATE=1.265
 CAUSE_SLUG="jingle-jam"
 FUNDRAISER_SLUG="jingle-jam-2023"
 FUNDRAISER_PUBLIC_ID="8601c634-c8a4-4a90-ac64-da037c683b46"
@@ -51,6 +52,7 @@ YEAR=2023
 COLLECTIONS_AVAILABLE=100000
 DOLLAR_OFFSET=36023.96
 DONATION_DIFFERENCE=6787
+CONVERSION_RATE=1.265
 CAUSE_SLUG="jingle-jam"
 FUNDRAISER_SLUG="jingle-jam-2022"
 FUNDRAISER_PUBLIC_ID="50f4c6e9-3879-4dbf-bbbe-5f804bbb82d6"
